chore: rename & doc
