+++ CONFLICT_RESOLUTION.md
@@ -1,0 +1,75 @@
+# Conflict Resolution Summary
+
+## Problem
+
+Next.js detected a conflict between App Router and Pages Router:
+
+- `src/pages/index.ts` (barrel export file)
+- `src/app/page.tsx` (App Router home page)
+
+## Solution
+
+Implemented a hybrid approach that resolves the conflict:
+
+### ✅ What Was Fixed
+
+1. **Removed Conflicting File**
+   - Deleted `src/pages/index.ts` (was incorrectly named)
+   - Created `src/pages/_exports.ts` for barrel exports
+
+2. **Added Pages Router Configuration**
+   - `src/pages/_app.tsx` - Custom App component with layout
+   - `src/pages/_document.tsx` - Custom Document with HTML structure
+   - Both files replicate the styling from `src/app/layout.tsx`
+
+3. **Route Distribution**
+   - **Home page (`/`)** → `src/app/page.tsx` (App Router)
+   - **All other pages** → `src/pages/*` (Pages Router)
+
+### 📁 Current Structure
+
+```
+src/
+├── app/
+│   ├── layout.tsx      # App Router layout
+│   ├── page.tsx        # Home page (/)
+│   └── api/
+└── pages/
+    ├── _app.tsx        # Pages Router app wrapper
+    ├── _document.tsx   # HTML document structure
+    ├── _exports.ts     # Barrel exports (renamed from index.ts)
+    ├── login.tsx       # /login
+    ├── register.tsx    # /register
+    └── ... (all other pages)
+```
+
+### 🚀 Benefits
+
+- ✅ **No conflicts** between routing systems
+- ✅ **Consistent styling** across all pages
+- ✅ **Modern home page** using App Router
+- ✅ **All application pages** using Pages Router
+- ✅ **Easy imports** via `_exports.ts`
+
+### 💡 Usage
+
+**Import pages:**
+
+```tsx
+import { LoginPage, CoursesPage } from "@/pages/_exports";
+```
+
+**Access routes:**
+
+- `/` → App Router home page
+- `/login` → Pages Router login page
+- `/courses` → Pages Router courses page
+- All other routes → Pages Router
+
+### 🎯 Result
+
+The conflict is resolved and you can continue development with:
+
+- Home page managed by App Router
+- All other pages managed by Pages Router
+- Consistent layout and styling across the entire application
