+++ src/pages/_document.tsx
@@ -1,0 +1,19 @@
+import { Html, Head, Main, NextScript } from "next/document";
+
+export default function Document() {
+  return (
+    <Html lang="en">
+      <Head>
+        <title>DooDoung</title>
+        <meta
+          name="description"
+          content="Your trusted platform for discovering and reviewing educational courses"
+        />
+      </Head>
+      <body>
+        <Main />
+        <NextScript />
+      </body>
+    </Html>
+  );
+}
