+++ src/pages/course.tsx
