+++ src/pages/course/[courseld].tsx
