+++ src/pages/index.tsx
@@ -1,0 +1,9 @@
+import { DefaultLayout } from "@/components/globalComponents";
+
+export default function LandingPage() {
+  return (
+    <DefaultLayout>
+      <p>This is Landing page</p>
+    </DefaultLayout>
+  );
+}
